[test] merge JRuby's Time tests into one .rb
